Fix breaking name change straight --> piecewise

+++ openpnm/network/GenericNetwork.py
@@ -606,7 +606,7 @@ class GenericNetwork(Base, ModelsMixin):
         [ 0  1  2  5  6 25 26]
         >>> Ps = pn.find_neighbor_pores(pores=[0, 2], flatten=False)
         >>> print(Ps)
-        [array([ 1,  5, 25]), array([ 1,  3,  7, 27])]
+        [array([ 1,  5, 25], dtype=int64), array([ 1,  3,  7, 27], dtype=int64)]
         >>> Ps = pn.find_neighbor_pores(pores=[0, 2], mode='xnor')
         >>> print(Ps)
         [1]
@@ -684,7 +684,7 @@ class GenericNetwork(Base, ModelsMixin):
         [  0   1 100 101 200 201]
         >>> Ts = pn.find_neighbor_throats(pores=[0, 1], flatten=False)
         >>> print(Ts)
-        [array([  0, 100, 200]), array([  0,   1, 101, 201])]
+        [array([  0, 100, 200], dtype=int64), array([  0,   1, 101, 201], dtype=int64)]
 
         """
         pores = self._parse_indices(pores)
@@ -819,7 +819,7 @@ class GenericNetwork(Base, ModelsMixin):
         >>> pn = op.network.Cubic(shape=[3, 3, 3])
         >>> Ps = pn.find_nearby_pores(pores=[0, 1], r=1)
         >>> print(Ps)
-        [array([3, 9]), array([ 2,  4, 10])]
+        [array([3, 9], dtype=int64), array([ 2,  4, 10], dtype=int64)]
         >>> Ps = pn.find_nearby_pores(pores=[0, 1], r=0.5)
         >>> print(Ps)
         [array([], dtype=int64), array([], dtype=int64)]
